+++ examples/timer.rs
@@ -8,6 +8,8 @@ use pushrod::render::widget_config::{CONFIG_COLOR_SECONDARY, CONFIG_PROGRESS};
 use pushrod::widgets::progress_widget::*;
 use pushrod::widgets::timer_widget::*;
 use sdl2::pixels::Color;
+use std::any::Any;
+use pushrod::widgets::text_widget::TextWidget;
 
 pub fn main() {
     let sdl_context = sdl2::init().unwrap();
@@ -57,6 +59,8 @@ pub fn main() {
             .get_numeric(CONFIG_PROGRESS)
             + 1)
             % 100;
+
+        eprintln!("{}", _widgets[widget1_id].widget.borrow_mut().as_any().is::<ProgressWidget>());
 
         _widgets[widget1_id]
             .widget