+++ src/widgets/image_button_widget.rs
@@ -27,6 +27,7 @@ use crate::widgets::image_widget::ImageWidget;
 use crate::widgets::text_widget::{TextJustify, TextWidget};
 use sdl2::pixels::Color;
 use std::collections::HashMap;
+use std::any::Any;
 
 /// This is the callback type that is used when an `on_click` callback is triggered from this
 /// `Widget`.
@@ -192,6 +193,10 @@ impl Widget for ImageButtonWidget {
         }
 
         self.button_clicked_callback(_widgets, _button, _clicks, _state);
+    }
+
+    fn as_any(&mut self) -> &dyn Any {
+        self
     }
 
     default_widget_properties!();