+++ src/widgets/toggle_button_widget.rs
@@ -25,6 +25,7 @@ use sdl2::video::Window;
 use crate::widgets::text_widget::{TextJustify, TextWidget};
 use sdl2::pixels::Color;
 use std::collections::HashMap;
+use std::any::Any;
 
 /// This is the callback type that is used when an `on_toggle` callback is triggered from this
 /// `Widget`.
@@ -216,6 +217,10 @@ impl Widget for ToggleButtonWidget {
         }
 
         self.button_clicked_callback(_widgets, _button, _clicks, _state);
+    }
+
+    fn as_any(&mut self) -> &dyn Any {
+        self
     }
 
     default_widget_properties!();